requirements.txt
phidata 
python-dotenv
yfinance
duckduckgo-search
groq
openai
agno
pypdf
lancedb
sentence-transformers
tantivy